# parallax_v1_3_artifact_pack/contracts/openapi/parallax_api_v1_3.yaml
openapi: 3.1.0
info:
  title: Parallax API
  version: 1.3.0
  description: Canonical v1.3 API contract for Parallax, a temporal-first personal
    intelligence application with timing analytics and user-controlled context awareness.
servers:
- url: https://api.parallax.local
  description: Local/staging base URL
security:
- bearerAuth: []
paths:
...
  /v1/activities:
    post:
      summary: Create activity
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      summary: List activities
      operationId: listActivities
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
          maximum: 100
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/activities/{activity_id}/preflight-checks:
    get:
      summary: List preflight checks
      operationId: listPreflightChecks
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PreflightCheck'
    post:
      summary: Create preflight check
      operationId: createPreflightCheck
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePreflightCheckRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreflightCheck'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/resource-dependencies:
    get:
      summary: List activity resource dependencies
      operationId: listActivityResourceDependencies
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Resource dependencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceDependency'
  /v1/activities/{activity_id}/preflight-checks/{check_id}/decision:
    post:
      summary: Decide preflight check suggestion or lifecycle state
      operationId: decidePreflightCheck
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: check_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecidePreflightCheckRequest'
      responses:
        '200':
          description: Preflight check after decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreflightCheck'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    MutationEnvelope:
      type: object
      additionalProperties: false
      properties:
        client_mutation_id:
          type: string
          minLength: 1
        client_device_id:
          type: string
          minLength: 1
        client_timestamp:
          type: string
          format: date-time
        idempotency_key:
          type: string
          minLength: 1
        client_sequence:
          type:
          - integer
          - 'null'
          minimum: 0
      required:
      - client_mutation_id
      - client_device_id
      - client_timestamp
      - idempotency_key
    ApiError:
      type: object
      additionalProperties: false
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
        retryable:
          type: boolean
        docs_ref:
          type:
          - string
          - 'null'
      required:
      - error_code
      - message
      - request_id
      - retryable
...
    Activity:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
          minLength: 1
        canonical_key:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - active
          - archived
          - merged
        merged_into_activity_id:
          type:
          - string
          - 'null'
          format: uuid
        default_timing_mode:
          type: string
          enum:
          - estimate_only
          - whole_task
          - checkpointed
          - routine
          - calibration
          - passive
        privacy_class:
          type: string
          enum:
          - normal
          - sensitive
          - private
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - id
      - user_id
      - display_name
      - status
      - default_timing_mode
      - privacy_class
      - created_at
      - updated_at
...
    ResourceDependency:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        activity_id:
          type: string
          format: uuid
        resource_name:
          type: string
        required_state:
          type:
          - string
          - 'null'
        usual_location:
          type:
          - string
          - 'null'
        failure_count:
          type: integer
          minimum: 0
        median_delay_seconds:
          type:
          - integer
          - 'null'
          minimum: 0
        p80_delay_seconds:
          type:
          - integer
          - 'null'
          minimum: 0
        suggest_precheck:
          type: boolean
        last_failed_at:
          type:
          - string
          - 'null'
          format: date-time
        created_from_event_id:
          type:
          - string
          - 'null'
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - id
      - user_id
      - activity_id
      - resource_name
      - failure_count
      - suggest_precheck
      - created_at
      - updated_at
    PreflightCheck:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        activity_id:
          type: string
          format: uuid
        check_text:
          type: string
          minLength: 1
        state:
          type: string
          enum:
          - suggested
          - active
          - snoozed
          - hidden
          - retired
        source:
          type: string
          enum:
          - user_created
          - model_suggested
          - resource_dependency
          - checkpoint_pattern
        confidence:
          type:
          - number
          - 'null'
          minimum: 0
          maximum: 1
        failure_count:
          type: integer
          minimum: 0
        last_triggered_at:
          type:
          - string
          - 'null'
          format: date-time
        source_event_id:
          type:
          - string
          - 'null'
          format: uuid
        source_dependency_id:
          type:
          - string
          - 'null'
          format: uuid
        snoozed_until:
          type:
          - string
          - 'null'
          format: date-time
        evidence_count:
          type: integer
          minimum: 0
        evidence_summary:
          type:
          - string
          - 'null'
        last_decided_at:
          type:
          - string
          - 'null'
          format: date-time
        decision_reason:
          type:
          - string
          - 'null'
        created_at:
          type:
          - string
          - 'null'
          format: date-time
        updated_at:
          type:
          - string
          - 'null'
          format: date-time
      required:
      - id
      - user_id
      - activity_id
      - check_text
      - state
      - source
      - failure_count
      - evidence_count
...
    CreateActivityRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        display_name:
          type: string
          minLength: 1
        description:
          type:
          - string
          - 'null'
        default_timing_mode:
          type: string
          enum:
          - estimate_only
          - whole_task
          - checkpointed
          - routine
          - calibration
          - passive
          default: whole_task
        privacy_class:
          type: string
          enum:
          - normal
          - sensitive
          - private
          default: normal
      required:
      - mutation
      - display_name
...
    CreatePreflightCheckRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        check_text:
          type: string
          minLength: 1
        source:
          type: string
          enum:
          - user_created
          - model_suggested
          - resource_dependency
          - checkpoint_pattern
          default: user_created
      required:
      - mutation
      - check_text
    DecidePreflightCheckRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        decision:
          type: string
          enum:
          - accept
          - hide
          - snooze
          - retire
        snoozed_until:
          type:
          - string
          - 'null'
          format: date-time
        reason:
          type:
          - string
          - 'null'
      required:
      - mutation
      - decision
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT